requirements.txt
GitPython
docker
pypiwin32